@@ -15,8 +15,8 @@
  */
 package org.thingsboard.server.service.security.system;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -49,6 +49,7 @@ import org.thingsboard.server.dao.exception.DataValidationException;
 import org.thingsboard.server.dao.settings.AdminSettingsService;
 import org.thingsboard.server.dao.user.UserService;
 import org.thingsboard.server.dao.user.UserServiceImpl;
+import org.thingsboard.server.dao.util.mapping.JacksonUtil;
 import org.thingsboard.server.service.security.exception.UserPasswordExpiredException;
 import org.thingsboard.server.utils.MiscUtils;
 
@@ -64,8 +65,6 @@ import static org.thingsboard.server.common.data.CacheConstants.SECURITY_SETTING
 @Service
 @Slf4j
 public class DefaultSystemSecurityService implements SystemSecurityService {
-
-    private static final ObjectMapper objectMapper = new ObjectMapper();
 
     @Autowired
     private AdminSettingsService adminSettingsService;
@@ -89,7 +88,7 @@ public class DefaultSystemSecurityService implements SystemSecurityService {
         AdminSettings adminSettings = adminSettingsService.findAdminSettingsByKey(tenantId, "securitySettings");
         if (adminSettings != null) {
             try {
-                securitySettings = objectMapper.treeToValue(adminSettings.getJsonValue(), SecuritySettings.class);
+                securitySettings = JacksonUtil.convertValue(adminSettings.getJsonValue(), SecuritySettings.class);
             } catch (Exception e) {
                 throw new RuntimeException("Failed to load security settings!", e);
             }
@@ -109,10 +108,10 @@ public class DefaultSystemSecurityService implements SystemSecurityService {
             adminSettings = new AdminSettings();
             adminSettings.setKey("securitySettings");
         }
-        adminSettings.setJsonValue(objectMapper.valueToTree(securitySettings));
+        adminSettings.setJsonValue(JacksonUtil.valueToTree(securitySettings));
         AdminSettings savedAdminSettings = adminSettingsService.saveAdminSettings(tenantId, adminSettings);
         try {
-            return objectMapper.treeToValue(savedAdminSettings.getJsonValue(), SecuritySettings.class);
+            return JacksonUtil.convertValue(savedAdminSettings.getJsonValue(), SecuritySettings.class);
         } catch (Exception e) {
             throw new RuntimeException("Failed to load security settings!", e);
         }
@@ -189,7 +188,7 @@ public class DefaultSystemSecurityService implements SystemSecurityService {
             JsonNode additionalInfo = user.getAdditionalInfo();
             if (additionalInfo instanceof ObjectNode && additionalInfo.has(UserServiceImpl.USER_PASSWORD_HISTORY)) {
                 JsonNode userPasswordHistoryJson = additionalInfo.get(UserServiceImpl.USER_PASSWORD_HISTORY);
-                Map<String, String> userPasswordHistoryMap = objectMapper.convertValue(userPasswordHistoryJson, Map.class);
+                Map<String, String> userPasswordHistoryMap = JacksonUtil.convertValue(userPasswordHistoryJson, new TypeReference<>() {});
                 for (Map.Entry<String, String> entry : userPasswordHistoryMap.entrySet()) {
                     if (encoder.matches(password, entry.getValue()) && Long.parseLong(entry.getKey()) > passwordReuseFrequencyTs) {
                         throw new DataValidationException("Password was already used for the last " + passwordPolicy.getPasswordReuseFrequencyDays() + " days");
