@@ -18,6 +18,8 @@ package org.thingsboard.server.dao.sql.rule;
 import org.springframework.data.repository.CrudRepository;
 import org.thingsboard.server.dao.model.sql.RuleNodeEntity;
 
-public interface RuleNodeRepository extends CrudRepository<RuleNodeEntity, String> {
+import java.util.UUID;
+
+public interface RuleNodeRepository extends CrudRepository<RuleNodeEntity, UUID> {
 
 }
