@@ -16,6 +16,7 @@
 package org.thingsboard.server.dao.util.mapping;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
@@ -37,6 +38,15 @@ public class JacksonUtil {
                     + fromValue + " cannot be converted to " + toValueType, e);
         }
     }
+
+    public static <T> T convertValue(Object fromValue, TypeReference<T> toValueTypeRef) {
+        try {
+            return fromValue != null ? OBJECT_MAPPER.convertValue(fromValue, toValueTypeRef) : null;
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException("The given object value: "
+                    + fromValue + " cannot be converted to " + toValueTypeRef, e);
+        }
+   }
 
     public static <T> T fromString(String string, Class<T> clazz) {
         try {
@@ -72,7 +82,9 @@ public class JacksonUtil {
     }
 
     public static <T> T clone(T value) {
-        return fromString(toString(value), (Class<T>) value.getClass());
+        @SuppressWarnings("unchecked")
+        Class<T> valueClass = (Class<T>) value.getClass();
+        return fromString(toString(value), valueClass);
     }
 
     public static <T> JsonNode valueToTree(T value) {
