@@ -34,7 +34,7 @@ import java.util.Map;
 @ClasspathSuite.ClassnameFilters({"org.thingsboard.server.msa.*Test"})
 public class ContainerTestSuite {
 
-    private static DockerComposeContainer testContainer;
+    private static DockerComposeContainer<?> testContainer;
 
     @ClassRule
     public static ThingsBoardDbInstaller installTb = new ThingsBoardDbInstaller();
@@ -43,7 +43,7 @@ public class ContainerTestSuite {
     public static DockerComposeContainer getTestContainer() {
         if (testContainer == null) {
             boolean skipTailChildContainers = Boolean.valueOf(System.getProperty("blackBoxTests.skipTailChildContainers"));
-            testContainer = new DockerComposeContainer(
+            testContainer = new DockerComposeContainer<>(
                     new File("./../../docker/docker-compose.yml"),
                     new File("./../../docker/docker-compose.postgres.yml"),
                     new File("./../../docker/docker-compose.postgres.volumes.yml"),
