@@ -44,7 +44,7 @@ public class HttpClientTest extends AbstractContainerTest {
         restClient.login("tenant@thingsboard.org", "tenant");
 
         Device device = createDevice("http_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         WsClient wsClient = subscribeToWebSocket(device.getId(), "LATEST_TELEMETRY", CmdsType.TS_SUB_CMDS);
         ResponseEntity deviceTelemetryResponse = restClient.getRestTemplate()
@@ -73,7 +73,7 @@ public class HttpClientTest extends AbstractContainerTest {
         TB_TOKEN = restClient.getToken();
 
         Device device = createDevice("test");
-        String accessToken = restClient.getCredentials(device.getId()).getCredentialsId();
+        String accessToken = restClient.getDeviceCredentialsByDeviceId(device.getId()).get().getCredentialsId();
         assertNotNull(accessToken);
 
         ResponseEntity deviceSharedAttributes = restClient.getRestTemplate()
@@ -92,6 +92,7 @@ public class HttpClientTest extends AbstractContainerTest {
 
         TimeUnit.SECONDS.sleep(3);
 
+        @SuppressWarnings("deprecation")
         Optional<JsonNode> allOptional = restClient.getAttributes(accessToken, null, null);
         assertTrue(allOptional.isPresent());
 
@@ -101,6 +102,7 @@ public class HttpClientTest extends AbstractContainerTest {
         assertEquals(mapper.readTree(createPayload().toString()), all.get("shared"));
         assertEquals(mapper.readTree(createPayload().toString()), all.get("client"));
 
+        @SuppressWarnings("deprecation")
         Optional<JsonNode> sharedOptional = restClient.getAttributes(accessToken, null, "stringKey");
         assertTrue(sharedOptional.isPresent());
 
@@ -108,6 +110,7 @@ public class HttpClientTest extends AbstractContainerTest {
         assertEquals(shared.get("shared").get("stringKey"), mapper.readTree(createPayload().get("stringKey").toString()));
         assertFalse(shared.has("client"));
 
+        @SuppressWarnings("deprecation")
         Optional<JsonNode> clientOptional = restClient.getAttributes(accessToken, "longKey,stringKey", null);
         assertTrue(clientOptional.isPresent());
 
