@@ -62,7 +62,7 @@ public class MqttClientTest extends AbstractContainerTest {
     public void telemetryUpload() throws Exception {
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         WsClient wsClient = subscribeToWebSocket(device.getId(), "LATEST_TELEMETRY", CmdsType.TS_SUB_CMDS);
         MqttClient mqttClient = getMqttClient(deviceCredentials, null);
@@ -89,7 +89,7 @@ public class MqttClientTest extends AbstractContainerTest {
 
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         WsClient wsClient = subscribeToWebSocket(device.getId(), "LATEST_TELEMETRY", CmdsType.TS_SUB_CMDS);
         MqttClient mqttClient = getMqttClient(deviceCredentials, null);
@@ -113,7 +113,7 @@ public class MqttClientTest extends AbstractContainerTest {
     public void publishAttributeUpdateToServer() throws Exception {
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         WsClient wsClient = subscribeToWebSocket(device.getId(), "CLIENT_SCOPE", CmdsType.ATTR_SUB_CMDS);
         MqttMessageListener listener = new MqttMessageListener();
@@ -144,7 +144,7 @@ public class MqttClientTest extends AbstractContainerTest {
     public void requestAttributeValuesFromServer() throws Exception {
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         WsClient wsClient = subscribeToWebSocket(device.getId(), "CLIENT_SCOPE", CmdsType.ATTR_SUB_CMDS);
         MqttMessageListener listener = new MqttMessageListener();
@@ -204,7 +204,7 @@ public class MqttClientTest extends AbstractContainerTest {
     public void subscribeToAttributeUpdatesFromServer() throws Exception {
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         MqttMessageListener listener = new MqttMessageListener();
         MqttClient mqttClient = getMqttClient(deviceCredentials, listener);
@@ -250,7 +250,7 @@ public class MqttClientTest extends AbstractContainerTest {
     public void serverSideRpc() throws Exception {
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         MqttMessageListener listener = new MqttMessageListener();
         MqttClient mqttClient = getMqttClient(deviceCredentials, listener);
@@ -297,7 +297,7 @@ public class MqttClientTest extends AbstractContainerTest {
     public void clientSideRpc() throws Exception {
         restClient.login("tenant@thingsboard.org", "tenant");
         Device device = createDevice("mqtt_");
-        DeviceCredentials deviceCredentials = restClient.getCredentials(device.getId());
+        DeviceCredentials deviceCredentials = restClient.getDeviceCredentialsByDeviceId(device.getId()).get();
 
         MqttMessageListener listener = new MqttMessageListener();
         MqttClient mqttClient = getMqttClient(deviceCredentials, listener);
